# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/db": {
      "get": {
        "operationId": "SyncController_getSnapshot",
        "parameters": [
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "description": "Role reading synchronized UI state.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read the full in-memory runtime snapshot for frontend synchronization.",
        "tags": [
          "database snapshot"
        ]
      }
    },
    "/api/db/snapshot": {
      "put": {
        "operationId": "SyncController_replaceSnapshot",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role required for runtime snapshot synchronization.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Replace or update in-memory runtime snapshot for active user session.",
        "tags": [
          "database snapshot"
        ]
      }
    },
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full CRUD; COO can read staff.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full CRUD; COO can read staff.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard response format.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "summary": "Create user account.",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_find",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full CRUD; COO can read staff.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full CRUD; COO can read staff.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UsersController_replace",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full CRUD; COO can read staff.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User for full CRUD; COO can read staff.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/system/modules": {
      "get": {
        "operationId": "SystemController_modules",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User only for system health, module status, and settings.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "system"
        ]
      }
    },
    "/api/system/settings": {
      "get": {
        "operationId": "SystemController_getSettings",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User only for system health, module status, and settings.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Read global platform settings.",
        "tags": [
          "system"
        ]
      },
      "patch": {
        "operationId": "SystemController_updateSettings",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Super User only for system health, module status, and settings.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update global platform settings (maintenance mode, session timeout, etc.).",
        "tags": [
          "system"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rorizon Review-4 API",
    "description": "NestJS in-memory backend with RBAC role headers for Rorizon workflows.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role for RBAC authorization. Allowed values: Super User, COO, Manager, Analyst."
      }
    },
    "schemas": {
      "SnapshotDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "example": 6
          },
          "users": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "organizations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "departments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "submissions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "resourceCategories": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "units": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "resourceTypes": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "resourceUnitCompatibilities": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "factorSources": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "factorVersions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "emissionFactors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "evidences": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "importBatches": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "importErrors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "resourceRecords": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "impactCalculations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "impactResults": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "managerSubmissions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "submissionTracker": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "reports": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "alerts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "auditLogs": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "plans": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "addons": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "subscriptions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "systemMetrics": {
            "type": "object"
          },
          "modules": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "cooKpis": {
            "type": "object"
          },
          "chartData": {
            "type": "object"
          },
          "managerChartData": {
            "type": "object"
          },
          "managerDailyUsage": {
            "type": "object"
          },
          "globalSettings": {
            "type": "object"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 105
          },
          "name": {
            "type": "string",
            "example": "New Manager"
          },
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": [
              "Super User",
              "COO",
              "Manager",
              "Analyst"
            ]
          },
          "email": {
            "type": "string",
            "example": "manager@techcorp.com"
          },
          "password": {
            "type": "string",
            "example": "Default@123"
          },
          "department": {
            "type": "string",
            "example": "Operations"
          },
          "departmentId": {
            "type": "string",
            "example": "dept-ops"
          },
          "assignedDepartmentIds": {
            "example": [
              "dept-ops"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organizationId": {
            "type": "string",
            "example": "org-techcorp"
          },
          "phone": {
            "type": "string",
            "example": "5550198236"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Inactive",
              "Pending"
            ]
          },
          "lastLogin": {
            "type": "string",
            "example": "Apr 01, 2026, 09:00 AM"
          }
        },
        "required": [
          "name",
          "role",
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 105
          },
          "name": {
            "type": "string",
            "example": "New Manager"
          },
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": [
              "Super User",
              "COO",
              "Manager",
              "Analyst"
            ]
          },
          "email": {
            "type": "string",
            "example": "manager@techcorp.com"
          },
          "password": {
            "type": "string",
            "example": "Default@123"
          },
          "department": {
            "type": "string",
            "example": "Operations"
          },
          "departmentId": {
            "type": "string",
            "example": "dept-ops"
          },
          "assignedDepartmentIds": {
            "example": [
              "dept-ops"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "organizationId": {
            "type": "string",
            "example": "org-techcorp"
          },
          "phone": {
            "type": "string",
            "example": "5550198236"
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": [
              "Active",
              "Inactive",
              "Pending"
            ]
          },
          "lastLogin": {
            "type": "string",
            "example": "Apr 01, 2026, 09:00 AM"
          }
        }
      },
...
    }
  }
}